+++ src/packages/@ncigdc/components/PortalContainer.js
@@ -2,18 +2,10 @@
 
 import React from 'react';
 import { connect } from 'react-redux';
-import { Route, Switch } from 'react-router-dom';
+import { Route } from 'react-router-dom';
 import Cookies from 'js-cookie';
 import { compose, lifecycle } from 'recompose';
-
-import ProjectRoute from '@ncigdc/routes/ProjectRoute';
-import FileRoute from '@ncigdc/routes/FileRoute';
-import CaseRoute from '@ncigdc/routes/CaseRoute';
-import AnnotationRoute from '@ncigdc/routes/AnnotationRoute';
-import GeneRoute from '@ncigdc/routes/GeneRoute';
-import SSMRoute from '@ncigdc/routes/SSMRoute';
-import SmartSearchRoute from '@ncigdc/routes/SmartSearchRoute';
-
+import PortalRoutes from '@ncigdc/routes';
 import Head from '@ncigdc/components/Head';
 import Header from '@ncigdc/components/Header';
 import Footer from '@ncigdc/components/Footer';
@@ -21,7 +13,6 @@ import NotificationContainer from '@ncigdc/components/NotificationContainer';
 import RelayLoadingContainer from '@ncigdc/components/RelayLoadingContainer';
 import ProgressContainer from '@ncigdc/components/ProgressContainer';
 import ModalContainer from '@ncigdc/components/ModalContainer';
-import NotFound from '@ncigdc/components/NotFound';
 
 import withRouter from '@ncigdc/utils/withRouter';
 import { GlobalTooltip } from '@ncigdc/uikit/Tooltip';
@@ -29,9 +20,6 @@ import { GlobalTooltip } from '@ncigdc/uikit/Tooltip';
 import styled from '@ncigdc/theme/styled';
 import { setModal } from '@ncigdc/dux/modal';
 import FirstTimeModal from '@ncigdc/components/Modals/FirstTimeModal';
-import LoadableWithLoading from '@ncigdc/components/LoadableWithLoading';
-
-import * as ModernComponents from '@ncigdc/modern_components';
 
 const SkipLink = styled.a({
   position: 'absolute',
@@ -41,30 +29,6 @@ const SkipLink = styled.a({
 });
 
 const FIRST_TIME_KEY = 'NCI-Warning';
-
-const HomeRoute = LoadableWithLoading({
-  loader: () => import('@ncigdc/routes/HomeRoute'),
-});
-
-const CartRoute = LoadableWithLoading({
-  loader: () => import('@ncigdc/routes/CartRoute'),
-});
-
-const RepositoryRoute = LoadableWithLoading({
-  loader: () => import('@ncigdc/routes/RepositoryRoute'),
-});
-
-const ExploreRoute = LoadableWithLoading({
-  loader: () => import('@ncigdc/routes/ExploreRoute'),
-});
-
-const ProjectsRoute = LoadableWithLoading({
-  loader: () => import('@ncigdc/routes/ProjectsRoute'),
-});
-
-const AnnotationsRoute = LoadableWithLoading({
-  loader: () => import('@ncigdc/routes/AnnotationsRoute'),
-});
 
 const enhance = compose(
   withRouter,
@@ -107,36 +71,10 @@ const PortalContainer = ({
         transition: 'padding 0.25s ease',
       }}
     >
-
       <Route
         children={p => <Head title={p.location.pathname.split('/')[1]} />}
       />
-      <Switch>
-        <Route exact path="/" component={HomeRoute} />
-        <Route exact path="/repository" component={RepositoryRoute} />
-        <Route exact path="/exploration" component={ExploreRoute} />
-        <Route exact path="/projects" component={ProjectsRoute} />
-        <Route exact path="/annotations" component={AnnotationsRoute} />
-        <Route exact path="/query" component={SmartSearchRoute} />
-        <Route path="/annotations/:id" component={AnnotationRoute} />
-        <Route path="/files/:id" component={FileRoute} />
-        {ProjectRoute}
-        {CaseRoute}
-        {AnnotationRoute}
-        {CartRoute}
-        {GeneRoute}
-        {SSMRoute}
-        <Route
-          path="/components/:component"
-          component={({ match, ...props }) => {
-            const Component = ModernComponents[match.params.component];
-            return Component
-              ? <Component />
-              : <h1>No matching component found.</h1>;
-          }}
-        />
-        <Route component={NotFound} />
-      </Switch>
+      <PortalRoutes />
     </div>
     <Footer />
